+++ modules/feedback/mobile/app/feedback.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { FeedbackScreen } from '../src/feedback/FeedbackScreen';
+
+export default function Feedback() {
+  return <FeedbackScreen />;
+}